+++ IterativeOptimizer.java
@@ -2,12 +2,12 @@
 public class IterativeOptimizer extends Optimizer {
     @Override
     public void optimize(UtilityCalculator utilityCalculator, MovePicker movePicker, AiEvaluator evaluator) {
-        int[] bestWeights = new int[utilityCalculator.getWeights().length];
+        double[] bestWeights = new double[utilityCalculator.getWeights().length];
         int testCount = 0;
         int bestPerformance = 0;
         for(int i = -10; i < 10; i++) {
             for(int j = -10; j < 10; j++) {
-                int[] weights = {i, j};
+                double[] weights = {i, j};
                 utilityCalculator.setWeights(weights);
                 int performance = evaluator.evaluateAveragePerformance(movePicker, utilityCalculator);
                 if(performance > bestPerformance) {
